requirements.txt
google-cloud-bigquery>=3.25.0
google-cloud-bigquery-storage>=2.24.0
db-dtypes>=1.2.0
pandas>=2.2.0
pyarrow>=16.0.0
python-dotenv>=1.0.1
typer>=0.12.0
rich>=13.7.0
matplotlib>=3.8.0
